[misc] Fix some admin-staff and meeting bugs.

+++ mrt/admin/staff.py
@@ -1,6 +1,5 @@
-from flask import render_template, flash
-from flask import request, redirect, url_for, jsonify
-from flask.ext.login import login_required
+from flask import flash, jsonify, redirect, render_template, request, url_for
+from flask.ext.login import login_required, current_user
 from flask.views import MethodView
 
 from mrt.admin.mixins import PermissionRequiredMixin
@@ -50,6 +49,9 @@ class StaffEdit(PermissionRequiredMixin, MethodView):
         staff = Staff.query.get_or_404(staff_id)
         role_user = RoleUser.query.filter_by(user=staff.user,
                                              meeting=None).one()
+        if staff.user == current_user:
+            message = 'You cannot delete your staff entry.'
+            return jsonify(status='error', message=message)
         db.session.delete(staff)
         db.session.delete(role_user)
         db.session.commit()
